--- client/app/components/cron/cronEditor.directive.js
@@ -1,9 +1,0 @@
-angular.module('directives').directive('cronEditor', () => ({
-  restrict: 'E',
-  replace: true,
-  scope: {
-    crontabObject: '=',
-  },
-  templateUrl: 'components/cron/cronEdit.html',
-  controller: 'cronEditorCtrl',
-}));
