--- client/app/components/cron/cronExpertHelp.directive.js
@@ -1,5 +1,0 @@
-angular.module('directives').directive('cronExpertHelp', () => ({
-  restrict: 'E',
-  replace: true,
-  templateUrl: 'components/cron/cronExpertHelp.html',
-}));
