--- client/app/components/cron/cronSimpleSummary.directive.js
@@ -1,9 +1,0 @@
-angular.module('directives').directive('cronSimpleSummary', () => ({
-  restrict: 'E',
-  replace: true,
-  scope: {
-    crontabObject: '=',
-  },
-  templateUrl: 'components/cron/cronSimpleSummary.html',
-  controller: 'cronSimpleSummaryCtrl',
-}));
