+++ client/app/hosting/cron/add-or-edit/hosting-cron-add-or-edit.html
@@ -118,9 +118,9 @@
         <div data-wizard-step
              data-wizard-step-valid="isValid(2)">
 
-            <cron-editor crontab-object="crontabObject"></cron-editor>
+            <wuc-cron-editor crontab-object="crontabObject"></wuc-cron-editor>
             <div data-wizard-step-help>
-                <cron-expert-help></cron-expert-help>
+                <wuc-cron-expert-help></wuc-cron-expert-help>
             </div>
         </div>
 
@@ -151,7 +151,7 @@
                         class="word-break"></dd>
                 </dl>
 
-                <cron-simple-summary crontab-object="crontabObject"></cron-simple-summary>
+                <wuc-cron-simple-summary crontab-object="crontabObject"></wuc-cron-simple-summary>
 
                 <dl class="dl-horizontal dl-lg">
                     <dt data-translate="hosting_tab_CRON_label_crontab"></dt>
