+++ client/app/webUniverseComponents/cron-validator/index.js
@@ -1,0 +1,12 @@
+import angular from 'angular';
+
+import WucCronValidator from './cron-validator';
+
+const moduleName = 'wucCronValidator';
+
+angular
+  .module(moduleName, [
+  ])
+  .service('WucCronValidator', WucCronValidator);
+
+export default moduleName;
