+++ client/app/webUniverseComponents/cron/cronEdit.html
@@ -1,24 +1,24 @@
 <div class="clearfix">
-    <p data-translate="hosting_tab_CRON_step2_title"></p>
+    <p data-translate="cron_step2_title"></p>
 
     <div class="row" data-ng-init="mode.active = 'simple'">
         <div class="col-md-6">
             <button class="btn btn-block btn-default btn-lg" data-ng-class="{'active': mode.active === 'simple'}"
-                    data-translate="hosting_tab_CRON_simple_mode"
+                    data-translate="cron_simple_mode"
                     data-ng-click="switchToSimpleMode()"></button>
         </div>
         <div class="col-md-6">
             <button class="btn btn-block btn-default btn-lg" data-ng-class="{'active': mode.active === 'expert'}"
-                    data-translate="hosting_tab_CRON_expert_mode"
+                    data-translate="cron_expert_mode"
                     data-ng-click="switchToExpertMode()"></button>
         </div>
     </div>
 
     <!-- MODE SIMPLE -->
     <div data-ng-show="mode.active === 'simple'"
-         data-ng-include="'components/cron/cronSimple.html'"></div>
+         data-ng-include="'wucCron_cronSimple.html'"></div>
 
     <!-- MODE EXPERT -->
     <div data-ng-show="mode.active === 'expert'"
-         data-ng-include="'components/cron/cronExpert.html'"></div>
+         data-ng-include="'wucCron_cronExpert.html'"></div>
 </div>
