+++ client/app/webUniverseComponents/cron/cronEditor.directive.js
@@ -1,0 +1,11 @@
+import template from './cronEdit.html';
+
+export default () => ({
+  restrict: 'E',
+  replace: true,
+  scope: {
+    crontabObject: '=',
+  },
+  template,
+  controller: 'wucCronEditorCtrl',
+});
