+++ client/app/webUniverseComponents/cron/cronExpertHelp.directive.js
@@ -1,0 +1,7 @@
+import template from './cronExpertHelp.html';
+
+export default () => ({
+  restrict: 'E',
+  replace: true,
+  template,
+});
