+++ client/app/webUniverseComponents/cron/cronExpertHelp.html
@@ -1,0 +1,19 @@
+<div>
+    <h3 data-translate="cron_expert_mode"></h3>
+    <ul class="list-unstyled">
+        <li><strong>M</strong>:&nbsp;<span data-translate="cron_step2_help_m"></span></li>
+        <li><strong>H</strong>:&nbsp;<span data-translate="cron_step2_help_h"></span></li>
+        <li><strong>DOM</strong>:&nbsp;<span data-translate="cron_step2_help_dom"></span></li>
+        <li><strong>MON</strong>:&nbsp;<span data-translate="cron_step2_help_mon" data-translate-values="{ t0: '&quot;JAN&quot;', t1: '&quot;DEC&quot;' }"></span></li>
+        <li><strong>DOW</strong>:&nbsp;<span data-translate="cron_step2_help_dow" data-translate-values="{ t0: '&quot;MON&quot;', t1: '&quot;SUN&quot;' }"></span></li>
+    </ul>
+    <h4 data-translate="cron_step2_help_subtitle1"></h4>
+    <ul class="list-unstyled">
+        <li><strong>*</strong>:&nbsp;<span data-translate="cron_step2_help_notation_1"></span></li>
+        <li><strong>2,4</strong>:&nbsp;<span data-translate="cron_step2_help_notation_2"></span></li>
+        <li><strong>4-7</strong>:&nbsp;<span data-translate="cron_step2_help_notation_3"></span></li>
+        <li><strong>*/3</strong>:&nbsp;<span data-translate="cron_step2_help_notation_4"></span></li>
+        <li><strong>10-20/3</strong>:&nbsp;<span data-translate="cron_step2_help_notation_5"></span></li>
+    </ul>
+    <p data-translate="cron_step2_help_subtitle2"></p>
+</div>
