+++ client/app/webUniverseComponents/cron/cronSimpleSummary.directive.js
@@ -1,0 +1,11 @@
+import template from './cronSimpleSummary.html';
+
+export default () => ({
+  restrict: 'E',
+  replace: true,
+  scope: {
+    crontabObject: '=',
+  },
+  template,
+  controller: 'wucCronSimpleSummaryCtrl',
+});
