+++ client/app/webUniverseComponents/cron/translations/Messages_fr_FR.xml
@@ -1,0 +1,81 @@
+<?xml version="1.0" encoding="utf-8"?>
+<translations>
+   <translation id="cron_dom_all" qtlid="154091">Tous les jours</translation>
+   <translation id="cron_dom_interval" qtlid="199429">Tous les N jours</translation>
+   <translation id="cron_dom_interval_value" qtlid="199442">Tous les {0} jours</translation>
+   <translation id="cron_dom_items" qtlid="61908">Jours fixes</translation>
+   <translation id="cron_dom_range" qtlid="199455">De tel jour à tel jour</translation>
+   <translation id="cron_dom_range_from" qtlid="88725">Du</translation>
+   <translation id="cron_dom_range_to" qtlid="35904">au</translation>
+   <translation id="cron_dow_all" qtlid="199468">Tous les jours de la semaine</translation>
+   <translation id="cron_dow_interval" qtlid="199494">Tous les N jours de la semaine</translation>
+   <translation id="cron_dow_interval_value" qtlid="199507">Tous les {0} jours de la semaine</translation>
+   <translation id="cron_dow_items" qtlid="199481">Jours de la semaine fixes</translation>
+   <translation id="cron_dow_range" qtlid="199520">De tel jour de la semaine à tel jour de la semaine</translation>
+   <translation id="cron_dow_range_from" qtlid="88725">Du</translation>
+   <translation id="cron_dow_range_to" qtlid="35904">au</translation>
+   <translation id="cron_expert_mode" qtlid="64616">Mode expert</translation>
+   <translation id="cron_hours_all" qtlid="8782">Toutes les heures</translation>
+   <translation id="cron_hours_interval" qtlid="199390">Toutes les N heures</translation>
+   <translation id="cron_hours_interval_value" qtlid="199403">Toutes les {0} heures</translation>
+   <translation id="cron_hours_items" qtlid="199377">Heures fixes</translation>
+   <translation id="cron_hours_range" qtlid="199416">De telle heure à telle heure</translation>
+   <translation id="cron_hours_range_from" qtlid="69077">De</translation>
+   <translation id="cron_hours_range_to" qtlid="8890">à</translation>
+   <translation id="cron_hours_value" qtlid="199611">{0} h</translation>
+   <translation id="cron_interval_pre_f" qtlid="199598">Toutes les</translation>
+   <translation id="cron_interval_pre_m" qtlid="62304">Tous les</translation>
+   <translation id="cron_label_days" qtlid="73505">Jours</translation>
+   <translation id="cron_label_days_min" qtlid="44260">jours</translation>
+   <translation id="cron_label_dow" qtlid="8854">Jours de la semaine</translation>
+   <translation id="cron_label_dow_min" qtlid="199364">jours de la semaine</translation>
+   <translation id="cron_label_hours" qtlid="8842">Heures</translation>
+   <translation id="cron_label_hours_min" qtlid="44236">heures</translation>
+   <translation id="cron_label_months" qtlid="37803">Mois</translation>
+   <translation id="cron_label_months_min" qtlid="40222">mois</translation>
+   <translation id="cron_mon_all" qtlid="199533">Tous les mois</translation>
+   <translation id="cron_mon_interval" qtlid="199559">Tous les N mois</translation>
+   <translation id="cron_mon_interval_value" qtlid="199572">Tous les {0} mois</translation>
+   <translation id="cron_mon_items" qtlid="199546">Mois fixes</translation>
+   <translation id="cron_mon_range" qtlid="199585">De tel mois à tel mois</translation>
+   <translation id="cron_mon_range_from" qtlid="69077">De</translation>
+   <translation id="cron_mon_range_to" qtlid="8890">à</translation>
+   <translation id="cron_simple_mode" qtlid="102849">Mode simple</translation>
+
+   <translation id="cron_step2_title" qtlid="199195">Sélectionnez la périodicité pour cette tâche :</translation>
+   <translation id="cron_step2_expert_help" qtlid="199208">Définissez la période, au format "Crontab", en insérant des valeurs pour l'heure ("{0}"), le jour ("{1}"), le mois ("{2}"), et le jour de la semaine ("{3}").</translation>
+   <translation id="cron_step2_expert_toocomplicated" qtlid="199221">Les valeurs sont trop complexes pour passer en mode simple.</translation>
+   <translation id="cron_step2_help_dom" qtlid="199689">Jour du mois (de 1 à 31)</translation>
+   <translation id="cron_step2_help_dow" qtlid="263038">Jour de la semaine (de 0 à 7, ou de {0} à {1})</translation>
+   <translation id="cron_step2_help_h" qtlid="199676">Heure (de 0 à 23)</translation>
+   <translation id="cron_step2_help_m" qtlid="199663">Minute (de 0 à 59)</translation>
+   <translation id="cron_step2_help_mon" qtlid="199702">Mois (de 1 à 12, ou de {0} à {1})</translation>
+   <translation id="cron_step2_help_notation_1" qtlid="199741">toutes les unités</translation>
+   <translation id="cron_step2_help_notation_2" qtlid="199754">uniquement le 2 et le 4</translation>
+   <translation id="cron_step2_help_notation_3" qtlid="199767">les unités entre 4 et 7 (4,5,6,7)</translation>
+   <translation id="cron_step2_help_notation_4" qtlid="199780">toutes les 3 unités (0,3,6,..)</translation>
+   <translation id="cron_step2_help_notation_5" qtlid="199793">toutes les 3 unités, mais uniquement entre le 10 et le 20 (10,13,16,19)</translation>
+   <translation id="cron_step2_help_subtitle1" qtlid="199728">Exemples de notation :</translation>
+   <translation id="cron_step2_help_subtitle2" qtlid="199806">Les minutes sont ignorées.</translation>
+
+   <translation id="cron_januar" qtlid="38745">Janvier</translation>
+   <translation id="cron_februar" qtlid="38229">Février</translation>
+   <translation id="cron_march" qtlid="38985">Mars</translation>
+   <translation id="cron_april" qtlid="38181">Avril</translation>
+   <translation id="cron_may" qtlid="38421">Mai</translation>
+   <translation id="cron_june" qtlid="38013">Juin</translation>
+   <translation id="cron_jully" qtlid="38085">Juillet</translation>
+   <translation id="cron_august" qtlid="38961">Août</translation>
+   <translation id="cron_september" qtlid="38409">Septembre</translation>
+   <translation id="cron_october" qtlid="39093">Octobre</translation>
+   <translation id="cron_november" qtlid="38169">Novembre</translation>
+   <translation id="cron_december" qtlid="37989">Décembre</translation>
+
+   <translation id="cron_monday" qtlid="6346">Lundi</translation>
+   <translation id="cron_tuesday" qtlid="6358">Mardi</translation>
+   <translation id="cron_wednesday" qtlid="6370">Mercredi</translation>
+   <translation id="cron_thursday" qtlid="6382">Jeudi</translation>
+   <translation id="cron_friday" qtlid="6394">Vendredi</translation>
+   <translation id="cron_saturday" qtlid="6406">Samedi</translation>
+   <translation id="cron_sunday" qtlid="6418">Dimanche</translation>
+</translations>
