+++ client/app/webUniverseComponents/index.js
@@ -1,12 +1,16 @@
 import angular from 'angular';
 
 import wucConverter from './converter';
+import wucCron from './cron';
+import wucCronValidator from './cron-validator';
 
 const moduleName = 'webUniverseComponents';
 
 angular
   .module(moduleName, [
     wucConverter,
+    wucCron,
+    wucCronValidator,
   ]);
 
 export default moduleName;
